Interface

QuantumAlgebra.BosonCreate

QuantumAlgebra.BosonDestroy

QuantumAlgebra.Corr

QuantumAlgebra.ExpVal

QuantumAlgebra.FermionCreate

QuantumAlgebra.FermionDestroy

QuantumAlgebra.OpSumAnalytic

QuantumAlgebra.param

QuantumAlgebra.scal

QuantumAlgebra.σ

QuantumAlgebra.σminus

QuantumAlgebra.σplus

QuantumAlgebra.Avac

QuantumAlgebra.CorrExpPerm_isless

QuantumAlgebra.CorrExpTup_isless

QuantumAlgebra.ascorr

QuantumAlgebra.ensure_compatible_sumind

QuantumAlgebra.extindices

QuantumAlgebra.prodcorr_inds

QuantumAlgebra.symmetric_index_nums

QuantumAlgebra.vacA

QuantumAlgebra.vacExpVal

QuantumAlgebra.σm

QuantumAlgebra.σp

QuantumAlgebra.σx

QuantumAlgebra.σy

QuantumAlgebra.σz

QuantumAlgebra.@boson_ops

QuantumAlgebra.@fermion_ops

QuantumAlgebra.BosonCreate — Type

QuantumAlgebra.BosonCreate(name,inds): represent bosonic creation operator $name_{inds}$

QuantumAlgebra.BosonDestroy — Type

QuantumAlgebra.BosonDestroy(name,inds): represent bosonic annihilation operator $name_{inds}$

QuantumAlgebra.Corr — Type

Corr(A::Operator)

Represents correlations of operator $A$. $A$ should be a product for this to make sense, in which case $⟨AB⟩_c = ⟨AB⟩ - ⟨A⟩⟨B⟩$, with corresponding extensions for products of more operators.

See also: ascorr

QuantumAlgebra.ExpVal — Type

ExpVal(A::Operator): represent expectation value $⟨A⟩$

QuantumAlgebra.FermionCreate — Type

QuantumAlgebra.FermionCreate(name,inds): represent fermionic creation operator $name_{inds}$

QuantumAlgebra.FermionDestroy — Type

QuantumAlgebra.FermionDestroy(name,inds): represent fermionic annihilation operator $name_{inds}$

QuantumAlgebra.OpSumAnalytic — Type

OpSumAnalytic(i::Symbol,A::Operator) or ∑(i,A): represent $\sum_{i} A$, with all possible values of $i$ assumed to be included

QuantumAlgebra.param — Type

param(g,(:i,:j),'n'): represent a scalar named parameter $g_{i,j}$. state can be purely real ('r'), not conjugated ('n'), or conjugated ('c')

QuantumAlgebra.scal — Type

Represent a scalar value (i.e., a number)

QuantumAlgebra.σ — Type

σ(a,inds): represent Pauli matrix $σ_{a,inds}$ for two-level system (TLS), where $a ∈ \{x,y,z\}$.

QuantumAlgebra.σminus — Type

QuantumAlgebra.σminus(inds): represent TLS annihilation operator $σ^-_{inds}$

QuantumAlgebra.σplus — Type

QuantumAlgebra.σplus(inds): represent TLS creation operator $σ^+_{inds}$

QuantumAlgebra.Avac — Function

Avac(A::Operator), vacA(A::Operator)

Simplify operator by assuming it is applied to the vacuum from the left or right, respectively. To be precise, Avac(A) returns $A'$ such that $A'|0⟩ = A|0⟩$, while vacA(A) does the same for $⟨0|A$.

QuantumAlgebra.CorrExpPerm_isless — Method

CorrExpPerm_isless(a,b)

isless for Tuples of Tuples representing products of Corr and ExpVal (see above) of a permutation of operators. E.g., ((1,3),(2,)) represents <A1 A3>_C <A2>. It is assumed that the total number of operators in a and b is equal, i.e., that sum(length.(a)) == sum(length.(b)).

QuantumAlgebra.CorrExpTup_isless — Method

CorrExpTup_isless(a,b)

isless for Tuples of integers that represent ExpVals and Corrs of sorted Operators (with n representing An such that n<m == An<Am). (n,) ⧋ ExpVal(An) (n,m,...) ⧋ Corr(AnAm...) Defined in such a way that the same order is obtained as with Operator objects

QuantumAlgebra.ascorr — Function

ascorr(expr::Operator)

Take an expression expr=A B C + D E... and write its expectation value in terms of single-body expectation values $⟨A⟩, ⟨B⟩, \ldots$, and many-body correlations $⟨AB⟩_c, ⟨ABC⟩_c$, etc. Currently, up to fourth-order correlations (i.e., products of four operators) are supported.

E.g., ascorr(adag(:n)*a(:n)) returns $⟨a^\dagger_n a_n⟩_c + ⟨a^\dagger_n⟩ ⟨a_n⟩$ (which is equal to $⟨a^\dagger_n a_n⟩$), while ascorr(adag(:n)*a(:m)*a(:n)) returns $\langle a_{n}^\dagger a_{m} a_{n} \rangle_{c} + \langle a_{n}^\dagger \rangle \langle a_{m} \rangle \langle a_{n} \rangle + \langle a_{n}^\dagger \rangle \langle a_{m} a_{n} \rangle_{c} + \langle a_{m} \rangle \langle a_{n}^\dagger a_{n} \rangle_{c} + \langle a_{n} \rangle \langle a_{n}^\dagger a_{m} \rangle_{c}$.

See also: ExpVal, Corr

QuantumAlgebra.ensure_compatible_sumind — Method

sum indices have no semantic meaning, so rename them in case they happen to occur in the other expression

QuantumAlgebra.extindices — Method

extindices(A::Operator) return externally visible indices of an expression

QuantumAlgebra.prodcorr_inds — Method

prodcorr_inds(N::Int)

for N operators, create an array of tuples of tuples that represents the terms in a sum of products of expectation values and correlators. Each tuple corresponds to a sum term, see CorrExpTup_isless and CorrExpPerm_isless for details of the format. The returned array and terms are sorted such that if the N operators are sorted, the represented expression is also sorted with the conventions of the QuantumAlgebra package. This allows to directly construct the nested OpSums and OpProds without having to go through simplification.

QuantumAlgebra.symmetric_index_nums — Method

symmetric_index_nums(A::Operator) return sequence of numbers of exchange-symmetric indices

QuantumAlgebra.vacA — Function

Avac(A::Operator), vacA(A::Operator)

Simplify operator by assuming it is applied to the vacuum from the left or right, respectively. To be precise, Avac(A) returns $A'$ such that $A'|0⟩ = A|0⟩$, while vacA(A) does the same for $⟨0|A$.

QuantumAlgebra.vacExpVal — Function

vacExpVal(A::Operator,S::Operator=scal(1))

Calculate the vacuum expectation value $⟨0|S^\dagger A S|0⟩$, i.e., the expectation value $⟨ψ|A|ψ⟩$ for the state defined by $|ψ⟩= S|0⟩$`.

QuantumAlgebra.σm — Method

σm(n): construct $σ^-_n = \frac12 σ_{x,n} - \frac{i}{2} σ_{y,n}$

QuantumAlgebra.σp — Method

σp(n): construct $σ^+_n = \frac12 σ_{x,n} + \frac{i}{2} σ_{y,n}$

QuantumAlgebra.σx — Method

σx(n): construct $σ_{x,n}$

QuantumAlgebra.σy — Method

σy(n): construct $σ_{y,n}$

QuantumAlgebra.σz — Method

σz(n): construct $σ_{z,n}$

QuantumAlgebra.@boson_ops — Macro

@boson_ops name: return functions for creating bosonic annihilation and creation operators with name name (i.e., wrappers of BosonDestroy and BosonCreate)

QuantumAlgebra.@fermion_ops — Macro

@fermion_ops name: return functions for creating fermionic annihilation and creation operators with name name (i.e., wrappers of FermionDestroy and FermionCreate)