Interface

Interface

QuantumAlgebra.Corr

QuantumAlgebra.ExpVal

QuantumAlgebra.OpSumAnalytic

QuantumAlgebra.a

QuantumAlgebra.adag

QuantumAlgebra.f

QuantumAlgebra.fdag

QuantumAlgebra.param

QuantumAlgebra.scal

QuantumAlgebra.σ

QuantumAlgebra.σminus

QuantumAlgebra.σplus

QuantumAlgebra.Avac

QuantumAlgebra.ascorr

QuantumAlgebra.ensure_compatible_sumind

QuantumAlgebra.extindices

QuantumAlgebra.symmetric_index_nums

QuantumAlgebra.vacA

QuantumAlgebra.vacExpVal

QuantumAlgebra.σm

QuantumAlgebra.σp

QuantumAlgebra.σx

QuantumAlgebra.σy

QuantumAlgebra.σz

QuantumAlgebra.Corr — Type.

Corr(A::Operator)

Represents correlations of operator $A$. $A$ should be a product for this to make sense, in which case $⟨AB⟩_c = ⟨AB⟩ - ⟨A⟩⟨B⟩$, with corresponding extensions for products of more operators.

See also: ascorr

source

QuantumAlgebra.ExpVal — Type.

ExpVal(A::Operator): represent expectation value $⟨A⟩$

source

QuantumAlgebra.OpSumAnalytic — Type.

OpSumAnalytic(i::Symbol,A::Operator) or ∑(i,A): represent $\sum_{i} A$, with all possible values of $i$ assumed to be included

source

QuantumAlgebra.a — Type.

a(inds): represent bosonic annihilation operator $a_{inds}$

source

QuantumAlgebra.adag — Type.

adag(inds): represent bosonic creation operator $a^†_{inds}$

source

QuantumAlgebra.f — Type.

f(inds): represent fermionic annihilation operator $f_{inds}$

source

QuantumAlgebra.fdag — Type.

fdag(inds): represent fermionic creation operator $f^†_{inds}$

source

QuantumAlgebra.param — Type.

param(g,(:i,:j),'n'): represent a scalar named parameter $g_{i,j}$. state can be purely real ('r'), not conjugated ('n'), or conjugated ('c')

source

QuantumAlgebra.scal — Type.

Represent a scalar value (i.e., a number)

source

QuantumAlgebra.Avac — Function.

Avac(A::Operator), vacA(A::Operator)

Simplify operator by assuming it is applied to the vacuum from the left or right, respectively. To be precise, Avac(A) returns $A'$ such that $A'|0⟩ = A|0⟩$, while vacA(A) does the same for $⟨0|A$.

source

QuantumAlgebra.ascorr — Function.

ascorr(expr::Operator)

Take an expression expr=A B C + D E... and write its expectation value in terms of single-body expectation values $⟨A⟩, ⟨B⟩, \ldots$, and many-body correlations $⟨AB⟩_c, ⟨ABC⟩_c$, etc. Currently, up to fourth-order correlations (i.e., products of four operators) are supported.

E.g., ascorr(adag(:n)*a(:n)) returns $⟨a^\dagger_n a_n⟩_c + ⟨a^\dagger_n⟩ ⟨a_n⟩$ (which is equal to $⟨a^\dagger_n a_n⟩$), while ascorr(adag(:n)*a(:m)*a(:n)) returns $\langle a_{n}^\dagger a_{m} a_{n} \rangle_{c} + \langle a_{n}^\dagger \rangle \langle a_{m} \rangle \langle a_{n} \rangle + \langle a_{n}^\dagger \rangle \langle a_{m} a_{n} \rangle_{c} + \langle a_{m} \rangle \langle a_{n}^\dagger a_{n} \rangle_{c} + \langle a_{n} \rangle \langle a_{n}^\dagger a_{m} \rangle_{c}$.

See also: ExpVal, Corr

source

QuantumAlgebra.vacA — Function.

Avac(A::Operator), vacA(A::Operator)

Simplify operator by assuming it is applied to the vacuum from the left or right, respectively. To be precise, Avac(A) returns $A'$ such that $A'|0⟩ = A|0⟩$, while vacA(A) does the same for $⟨0|A$.

source

QuantumAlgebra.vacExpVal — Function.

vacExpVal(A::Operator,S::Operator=scal(1))

Calculate the vacuum expectation value $⟨0|S^\dagger A S|0⟩$, i.e., the expectation value $⟨ψ|A|ψ⟩$ for the state defined by $|ψ⟩= S|0⟩$`.

source

QuantumAlgebra.σm — Method.

σm(n): construct $σ^-_n = \frac12 σ_{x,n} - \frac{i}{2} σ_{y,n}$

source

QuantumAlgebra.σp — Method.

σp(n): construct $σ^+_n = \frac12 σ_{x,n} + \frac{i}{2} σ_{y,n}$

source

QuantumAlgebra.σx — Method.

σx(n): construct $σ_{x,n}$

source

QuantumAlgebra.σy — Method.

σy(n): construct $σ_{y,n}$

source

QuantumAlgebra.σz — Method.

σz(n): construct $σ_{z,n}$

source

QuantumAlgebra.σ — Type.

σ(a,n): represent Pauli matrix $σ_{a,n}$ for two-level system (TLS) $n$, where $a ∈ \{x,y,z\}$ or $\{1,2,3\}$ is the type of Pauli matrix.

source

QuantumAlgebra.σminus — Type.

QuantumAlgebra.σminus(inds): represent TLS annihilation operator $σ^-_{inds}$

source

QuantumAlgebra.σplus — Type.

QuantumAlgebra.σplus(inds): represent TLS creation operator $σ^+_{inds}$

source

QuantumAlgebra.ensure_compatible_sumind — Method.

sum indices have no semantic meaning, so rename them in case they happen to occur in the other expression

source

QuantumAlgebra.extindices — Method.

extindices(A::Operator) return externally visible indices of an expression

source

QuantumAlgebra.symmetric_index_nums — Method.

symmetric_index_nums(A::Operator) return sequence of numbers of exchange-symmetric indices

source